requirements.txt

<old_str>
django-routers>=0.2
fastapi>=0.115.13
httpx>=0.25.0
passlib>=1.7.4
psycopg2-binary>=2.9.0
pydantic>=2.11.7
pyjwt>=2.10.1
pytest>=8.4.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
requests>=2.31.0
sqlalchemy>=2.0.41
uvicorn>=0.34.3
email-validator>=2.1.0
</old_str>
<new_str>
fastapi>=0.115.13
uvicorn>=0.34.3
sqlalchemy>=2.0.41
psycopg2-binary>=2.9.0
pydantic[email]>=2.11.7
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
requests>=2.31.0
httpx>=0.25.0
pytest>=8.4.1
pytest-asyncio>=0.21.0
email-validator>=2.1.0
</new_str>
